requirements.txt
# Core dependencies
mcp>=1.0.0
pydantic>=2.0.0
google-generativeai>=0.3.0

# Data processing
aiofiles>=23.0.0
pathspec>=0.11.0
jinja2>=3.1.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Logging and utilities  
structlog>=23.0.0
python-dotenv>=1.0.0